src/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
azqueuetweeter==0.1.1
    # via -r requirements.in
azure-core==1.26.1
    # via
    #   azure-storage-queue
    #   msrest
azure-storage-queue==12.5.0
    # via
    #   -r requirements.in
    #   azqueuetweeter
certifi==2022.9.24
    # via
    #   msrest
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via flask
cryptography==38.0.3
    # via azure-storage-queue
flask==2.2.2
    # via
    #   -r requirements.in
    #   flask-wtf
flask-wtf==1.0.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
isodate==0.6.1
    # via msrest
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
    #   wtforms
msrest==0.7.1
    # via azure-storage-queue
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   tweepy
pycparser==2.21
    # via cffi
python-dotenv==0.21.0
    # via -r requirements.in
requests==2.28.1
    # via
    #   azure-core
    #   msrest
    #   requests-oauthlib
    #   tweepy
requests-oauthlib==1.3.1
    # via
    #   msrest
    #   tweepy
six==1.16.0
    # via
    #   azure-core
    #   isodate
tweepy==4.12.1
    # via azqueuetweeter
typing-extensions==4.4.0
    # via azure-core
urllib3==1.26.12
    # via requests
werkzeug==2.2.2
    # via flask
wtforms==3.0.1
    # via
    #   -r requirements.in
    #   flask-wtf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
